inbox pattern support
by setting the `:inbox_table` and `inbox_column`, and setting up the
inbox table, one sets transparent support for the inbox pattern. on each
transaction for event handling, an INSERT query is performed with the
unique id. if it fails because of unique id, the event is discarded.

At least in postgres, if tuw INSERT ON CONFLICT are queried at the same
time, with the same id, the last one awaits for the first one to finish.
This may not be ideal if the handlers are heavy.